scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.8.0",
    "decorest>=0.1.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "sermonaudio>=6.15.2",
]
